Add backwards compatibility for wp_check_filetype
This adds a backward compat shim for cases where a mimes array is passed as the second parameter to the function and then attempts to pass the file to `wp_check_file_mime` which has an additional `wp_check_ext_mime` guesser added which will return a mime type based on the old extension look up method so we can continue to support non-path values.

Renames all mime guesser functions and filters from wp_get_* to wp_check_* to better reflect that these aren't simply getter functions, but are checking the value of the file being passed in.

Updates tests.